test: implement a lightweight sanity e2e not using a test runner. added playwright to repo. still low number of deps.
